Extract kitty CSI-u keyboard decoder
