+++ src/Termina/Input/EscapeSequenceParser.cs
@@ -207,7 +207,7 @@ internal sealed class EscapeSequenceParser
                 // CSI u (kitty keyboard protocol): ESC[keycode;modifiersu
                 if (key.KeyChar == 'u' && seq.Length >= 2)
                 {
-                    if (TryParseCsiU(seq, out var csiEvent) && csiEvent is not null)
+                    if (KittyCsiUKeyboardDecoder.TryDecode(seq, out var csiEvent) && csiEvent is not null)
                     {
                         TerminaTrace.Input.Debug(this, "ESP: CSI u detected: {0}", seq);
                         results.Add(csiEvent);
@@ -369,126 +369,4 @@ internal sealed class EscapeSequenceParser
         return false;
     }
 
-    /// <summary>
-    /// Attempts to parse a CSI u (kitty keyboard protocol) sequence.
-    /// Format: <c>[keycode;modifiersu</c> where keycode is a Unicode codepoint
-    /// and modifiers is a 1-based bitmask (shift=2, alt=3, ctrl=5, etc.).
-    /// See <see href="https://sw.kovidgoyal.net/kitty/keyboard-protocol/"/> for the full specification.
-    /// </summary>
-    /// <param name="seq">The buffered sequence string, e.g. <c>[13;5u</c>.</param>
-    /// <param name="result">The resulting <see cref="KeyPressed"/> event, if parsing succeeded.</param>
-    /// <returns><c>true</c> if the sequence was successfully parsed.</returns>
-    private static bool TryParseCsiU(string seq, out KeyPressed? result)
-    {
-        result = null;
-
-        // seq = "[keycode[;modifiers[:event][;text...]]u" — strip leading "[" and trailing "u"
-        if (seq.Length < 3 || seq[0] != '[' || seq[^1] != 'u')
-            return false;
-
-        var inner = seq[1..^1];
-        var semicolon = inner.IndexOf(';');
-
-        int keycode;
-        var modValue = 1; // default: no modifiers (1-based)
-
-        if (semicolon < 0)
-        {
-            // No modifier field — just keycode (possibly with :alternate sub-fields we ignore).
-            var keyPart = inner;
-            var sub = keyPart.IndexOf(':');
-            if (sub >= 0) keyPart = keyPart[..sub];
-            if (!int.TryParse(keyPart, out keycode))
-                return false;
-        }
-        else
-        {
-            var keyPart = inner[..semicolon];
-            var subKey = keyPart.IndexOf(':');
-            if (subKey >= 0) keyPart = keyPart[..subKey];
-            if (!int.TryParse(keyPart, out keycode))
-                return false;
-
-            // Modifier field may have a :<event-type> sub-field, and a trailing ;<text-codepoints>.
-            var rest = inner[(semicolon + 1)..];
-            // Strip any trailing ;text-codepoints (we ignore associated text in the prototype).
-            var nextSemi = rest.IndexOf(';');
-            if (nextSemi >= 0) rest = rest[..nextSemi];
-
-            var modPart = rest;
-            var colon = rest.IndexOf(':');
-            if (colon >= 0)
-            {
-                modPart = rest[..colon];
-                if (!int.TryParse(rest[(colon + 1)..], out var eventType))
-                    return false;
-                if (eventType != 1) return true; // swallow repeat / release
-            }
-            if (!int.TryParse(modPart, out modValue) || modValue < 1) return false;
-        }
-
-        // Swallow modifier-key-alone press events (left/right Shift/Ctrl/Alt/Super/Hyper/Meta,
-        // CapsLock, NumLock, etc.) — the kitty PUA range 57441..57454 carries these and they
-        // overwhelm the prototype's input loop with noise on every keystroke.
-        if (keycode is >= 57441 and <= 57454)
-            return true; // parsed-and-swallowed
-
-        // Decode 1-based modifier bitmask: value = 1 + (shift ? 1 : 0) + (alt ? 2 : 0) + (ctrl ? 4 : 0)
-        var modBits = modValue - 1;
-        var shift = (modBits & 1) != 0;
-        var alt = (modBits & 2) != 0;
-        var ctrl = (modBits & 4) != 0;
-
-        var (consoleKey, keyChar) = MapKeycodeToConsoleKey(keycode);
-        // Modifier-alone or unknown keys we don't want to surface get ConsoleKey.None and '\0' —
-        // swallow them too rather than surfacing a confusing empty KeyPressed.
-        if (consoleKey == ConsoleKey.None && keyChar == '\0')
-            return true;
-
-        result = new KeyPressed(new ConsoleKeyInfo(keyChar, consoleKey, shift, alt, ctrl));
-        return true;
-    }
-
-    /// <summary>
-    /// Maps a kitty keyboard protocol keycode (Unicode codepoint or functional key code)
-    /// to a <see cref="ConsoleKey"/> and character.
-    /// </summary>
-    private static (ConsoleKey Key, char Char) MapKeycodeToConsoleKey(int keycode) => keycode switch
-    {
-        13 => (ConsoleKey.Enter, '\r'),
-        9 => (ConsoleKey.Tab, '\t'),
-        127 => (ConsoleKey.Backspace, '\b'),
-        27 => (ConsoleKey.Escape, '\x1b'),
-        // Printable ASCII range
-        >= 32 and <= 126 => (CharToConsoleKey((char)keycode), (char)keycode),
-        // Kitty functional-key PUA codepoints (https://sw.kovidgoyal.net/kitty/keyboard-protocol/#functional)
-        57344 => (ConsoleKey.Escape, '\x1b'),
-        57345 => (ConsoleKey.Enter, '\r'),
-        57346 => (ConsoleKey.Tab, '\t'),
-        57347 => (ConsoleKey.Backspace, '\b'),
-        57348 => (ConsoleKey.Insert, '\0'),
-        57349 => (ConsoleKey.Delete, '\0'),
-        57350 => (ConsoleKey.LeftArrow, '\0'),
-        57351 => (ConsoleKey.RightArrow, '\0'),
-        57352 => (ConsoleKey.UpArrow, '\0'),
-        57353 => (ConsoleKey.DownArrow, '\0'),
-        57354 => (ConsoleKey.PageUp, '\0'),
-        57355 => (ConsoleKey.PageDown, '\0'),
-        57356 => (ConsoleKey.Home, '\0'),
-        57357 => (ConsoleKey.End, '\0'),
-        57358 => (ConsoleKey.None, '\0'), // CapsLock — no ConsoleKey constant; swallow.
-        // F1..F12 (Kitty PUA 57364..57375)
-        >= 57364 and <= 57375 => (ConsoleKey.F1 + (keycode - 57364), '\0'),
-        // Default: use the codepoint as the char, no specific ConsoleKey
-        _ => (ConsoleKey.None, keycode < 128 ? (char)keycode : '\0')
-    };
-
-    private static ConsoleKey CharToConsoleKey(char c) => c switch
-    {
-        >= 'a' and <= 'z' => ConsoleKey.A + (c - 'a'),
-        >= 'A' and <= 'Z' => ConsoleKey.A + (c - 'A'),
-        >= '0' and <= '9' => ConsoleKey.D0 + (c - '0'),
-        ' ' => ConsoleKey.Spacebar,
-        _ => ConsoleKey.None
-    };
 }
